Pull locale logic into separate file. Use locale when formatting dates
